+++ stroom-dashboard/stroom-dashboard-client/src/main/java/stroom/dashboard/client/table/TableSettingsPresenter.java
@@ -19,13 +19,16 @@ package stroom.dashboard.client.table;
 import com.google.inject.Inject;
 import com.google.web.bindery.event.shared.EventBus;
 import stroom.dashboard.client.main.SettingsPresenter;
+import stroom.dashboard.client.table.cf.RulesPresenter;
 import stroom.widget.tab.client.presenter.LinkTabsLayoutView;
 
 public class TableSettingsPresenter extends SettingsPresenter {
     @Inject
     public TableSettingsPresenter(final EventBus eventBus, final LinkTabsLayoutView view,
-                                  final BasicTableSettingsPresenter basicSettingsPresenter) {
+                                  final BasicTableSettingsPresenter basicSettingsPresenter,
+                                  final RulesPresenter rulesPresenter) {
         super(eventBus, view);
         addTab("Basic", basicSettingsPresenter);
+        addTab("Conditional Formatting", rulesPresenter);
     }
 }